tental refactor-not completed yet!
